Use Upholds= to ensure services run after daemon-reload
When a systemd-sysext image is loaded, normally a daemon-reload is done
but the targets are not evaluated, meaning that enabled services don't
start. In Flatcar we ship a workaround for that by restarting a certain
list of targets but this is not upstream behavior. Currently upstream
also doesn't do the daemon-reload but we can file an upstream PR for
that.