Update instructions to opt-out of a future Flatcar Docker sysext
Instances deployed with a custom Docker version shouldn't suddenly get
the default one of Flatcar when we ship it as sysext. The docs to opt-
out were assuming that they should mask the default sysext from /usr.
However, /usr as location for sysext images is removed and we have to
enable the Flatcar default sysext through a symlink in /etc.
Document that the user should create this symlink to prevent it from
popping up through an update.