Add docker compose CLI plugin sysext
The standalone docker-compose binary is replaced by a new CLI plugin.
This CLI plugin needs to live under /usr, thus we have to use a sysext
image.
Add a helper to create a sysext image. While we could also add it by
default to the docker sysext script, the separate usage is of more
interest because it can be used with Flatcar's default Docker.
Done by following
https://docs.docker.com/compose/install/linux/#install-the-plugin-manually