bake.sh: Set SOURCE_DATE_EPOCH if not defined
The binaries consumed from upstream should have the same checksums
already but the resulting squashfs does not because it contains the
time stamps of the build.
Define the SOURCE_DATE_EPOCH variable which is enough to make mksquashfs
reproducible. The user can pass an own value if there is a fixed
reference time. For semantics, see
https://reproducible-builds.org/docs/source-date-epoch/

Fixes https://github.com/flatcar/sysext-bakery/issues/25