Add extension overview and release all repackaged extensions
Recently more build scripts got added and not all are covered in the
release yet. Add a table that gives an overview on available extensions
and whether they are released or not. We try to release all extensions,
I didn't do it for keepalived now because it was quite slow to build
for arm64. While at it fix the README order a bit, added a Butane
template and made the release instructions more robust against mistakes.