release_build: Handle filenames with non-url characters
k3s and rke have '+' in their version number, which is part of the sysext
filename. The url mangles that into %2B and our script preserves that as the
filename of the next upload. This breaks uploading on the third run. Extract
the name of the sysext from the release assets list to fix.