Add Azure Keyvault tooling
This adds a script which sets up Azure keyvault with a signing key,
managed identity and federated credentials to allow GitHub Actions
running within the context of this repo to perform sign operations using
the keyvault. This also adds an action which pulls the certificate from
keyvault and publishes it as a release.