Add example extensions built with flix.sh and flatwrap.sh
To show how flix.sh and flatwrap.sh can be used to ship regular Linux
packages as sysext we can add two examples.
As flix example include an extension for the tilde text editor from
Debian and as flatwrap example include an extension for btop from
Alpine (originally I used htop but we might ship this in Flatcar later
by default - note that you need to set LANG=en_US.UTF-8 or run with btop
with --utf-force on Flatcar).

Signed-off-by: Kai Lueke <kailuke@microsoft.com>